Make defaults handling class property friendly
There's an open ECMAScript proposal to add value properties to
classes: https://gist.github.com/jeffmo/054df782c05639da2adb

With that syntax it's possible to turn this:

```js
class Client extends Gofer {}
Client.prototype.serviceName = 'xyz';
```

Into this:

```js
class Client extends Gofer {
  serviceName = 'xyz';
}
```

The problem is that the property initialization has to happen after
`this` is available. Which in the case of ECMAScript classes means
"super() was called". So if we want to allow the above syntax, we
may not use `serviceName` inside of the constructor.

The solution here is to delay calculating the defaults until first
use. By providing properties for `client.defaults`, this should not
break the interface.